feat: 전화번호 인증이 없을 때 정지 상태를 해제할 수 없도록 변경

+++ src/main/java/com/example/chook/admin/service/AdminServiceImpl.java
@@ -112,6 +112,8 @@ public class AdminServiceImpl implements AdminService {
     Member member = memberRepository.findById(memberId).orElseThrow(() ->
       new EntityNotFoundException("Member with id: " + memberId + " not found")
     );
+    if (!member.isPhoneVerified())
+      throw new IllegalStateException("휴대전화 번호가 인증된 상태여야만 정지를 해제할 수 있습니다.");
     boolean isSuspended = memberSuspensionRepository.existsById(memberId);
     if (!isSuspended) return false;
     memberSuspensionRepository.deleteById(memberId);
